removed display_indexes.lib.php. added tbl_display.lib.php and extracted methods from tbl_indexes.php to tbl_indexes.lib.php
